requirements.txt
streamlit>=1.54.0
google-genai>=0.8.0
plotly>=6.1.1
pandas==2.2.2
openpyxl>=3.1.0
pypdf>=5.0.0
kaleido>=1.2.0